requirements.txt
fastapi
uvicorn
numpy
Pillow
tensorflow
python-multipart
streamlit
requests
matplotlib
streamlit-drawable-canvas